+++ cpp/oneapi/dal/backend/primitives/distance/squared_l2_distance_auxilary.hpp
