+++ cpp/oneapi/dal/backend/primitives/distance/squared_l2_distance_dpc.cpp
@@ -15,100 +15,57 @@
 *******************************************************************************/
 
 #include "oneapi/dal/backend/primitives/distance/distance.hpp"
-#include "oneapi/dal/backend/primitives/distance/squared_l2_distance.hpp"
+#include "oneapi/dal/backend/primitives/distance/squared_l2_distance_auxilary.hpp"
 
 #include "oneapi/dal/backend/primitives/blas.hpp"
 #include "oneapi/dal/backend/primitives/reduction.hpp"
 
 namespace oneapi::dal::backend::primitives {
 
-template <typename Float>
-sycl::event compute_squared_l2_norms(sycl::queue& q,
-                                     const ndview<Float, 2>& inp,
-                                     ndview<Float, 1>& out,
-                                     const event_vector& deps) {
-    return reduce_by_rows(q, inp, out, sum<Float>{}, square<Float>{}, deps);
-}
-
-template <typename Float>
-std::tuple<array<Float>, sycl::event> compute_squared_l2_norms(sycl::queue& q,
-                                                               const ndview<Float, 2>& inp,
-                                                               const event_vector& deps,
-                                                               const sycl::usm::alloc& alloc) {
-    const auto n_samples = inp.get_dimension(0);
-    auto res_array = array<Float>::empty(q, n_samples, alloc);
-    auto res_wrapper = ndview<Float, 1>::wrap(res_array.get_mutable_data(), { n_samples });
-    return { res_array, compute_squared_l2_norms(q, inp, res_wrapper, deps) };
-}
-
-template <typename Float>
-sycl::event scatter_2d(sycl::queue& q,
-                       const ndview<Float, 1>& inp1,
-                       const ndview<Float, 1>& inp2,
-                       ndview<Float, 2>& out,
-                       const event_vector& deps) {
-    const auto out_stride = out.get_leading_stride();
-    const auto n_samples1 = inp1.get_dimension(0);
-    const auto n_samples2 = inp2.get_dimension(0);
-    ONEDAL_ASSERT(n_samples1 <= out.get_dimension(1));
-    ONEDAL_ASSERT(n_samples2 <= out.get_dimension(0));
-    const auto* inp1_ptr = inp1.get_data();
-    const auto* inp2_ptr = inp2.get_data();
-    auto* out_ptr = out.get_mutable_data();
-    sycl::range<2> out_range(n_samples1, n_samples2);
-    return q.submit([&](sycl::handler& h) {
+template<typename Float>
+sycl::event distance<Float, squared_l2_metric<Float>>::initialize(const ndview<Float, 2>& inp1,
+                                                                  const ndview<Float, 2>& inp2,
+                                                                  const event_vector& deps) {
+    return q_.submit([&](sycl::handler& h) {
         h.depends_on(deps);
-        h.parallel_for(out_range, [=](sycl::id<2> idx) {
-            auto& out_place = *(out_ptr + out_stride * idx[0] + idx[1]);
-            out_place = inp1_ptr[idx[0]] + inp2_ptr[idx[1]];
-        });
     });
 }
 
-template <typename Float>
-sycl::event compute_inner_product(sycl::queue& q,
-                                  const ndview<Float, 2>& inp1,
-                                  const ndview<Float, 2>& inp2,
-                                  ndview<Float, 2>& out,
-                                  const event_vector& deps) {
-    check_inputs(inp1, inp2, out);
-    return gemm(q, inp1, inp2.t(), out, Float(-2.0), Float(+1.0), deps);
+template<typename Float>
+auto distance<Float, squared_l2_metric<Float>>::get_norms(const helper_ptr_t& helper,
+                                                          const ndview<Float, 2>& inp,
+                                                          const event_vector& deps) const -> norms_res_t {
+    // TODO: utilize helpers
+    auto [res_array, res_event] = compute_squared_l2_norms(q_, inp, deps);
+    return {res_array, res_event};
 }
 
-template <typename Float>
-sycl::event compute_squared_l2_distance(sycl::queue& q,
-                                        const ndview<Float, 2>& inp1,
-                                        const ndview<Float, 2>& inp2,
-                                        ndview<Float, 2>& out,
-                                        const event_vector& deps) {
-    check_inputs(inp1, inp2, out);
-    auto [l2_inp1_arr, l2_inp1_event] = compute_squared_l2_norms(q, inp1, deps);
-    auto [l2_inp2_arr, l2_inp2_event] = get_norms(q, inp2, deps);
-    auto scatter_event =
-        scatter_norms(q, l2_inp1_arr, l2_inp2_arr, out, { l2_inp1_event, l2_inp2_event });
-    return inner_product(q, inp1, inp2, out, { scatter_event });
+template<typename Float>
+sycl::event distance<Float, squared_l2_metric<Float>>::operator() (const ndview<Float, 2>& inp1,
+                                                                   const ndview<Float, 2>& inp2,
+                                                                   ndview<Float, 2>& out,
+                                                                   const event_vector& deps) const {
+    auto [norms1_array, norms1_event] = get_norms(helper1_, inp1, deps);
+    auto [norms2_array, norms2_event] = get_norms(helper2_, inp2, deps);
+    const auto norms1_view = ndview<Float, 1>::wrap(norms1_array.get_data(), { norms1_array.get_count() });
+    const auto norms2_view = ndview<Float, 1>::wrap(norms2_array.get_data(), { norms2_array.get_count() });
+    auto scatter_event = scatter_2d(q_, norms1_view, norms2_view, out, {norms1_event, norms2_event});
+    return compute_inner_product(q_, inp1, inp2, out, {scatter_event});
 }
+
+template<typename Float>
+class l2_helper {
+    Float val = 42;
+};
+
+template<typename Float>
+distance<Float, squared_l2_metric<Float>>::distance(sycl::queue& q) : 
+                                            q_{ q }, helper1_{ nullptr }, helper2_{ nullptr } {}
+
+template<typename Float>
+distance<Float, squared_l2_metric<Float>>::~distance()  = default;
 
 #define INSTANTIATE(F)                                                      \
-    template sycl::event compute_squared_l2_norms<F>(sycl::queue&,          \
-                                                     const ndview<F, 2>&,   \
-                                                     ndview<F, 1>&,         \
-                                                     const event_vector&);  \
-    template std::tuple<array<F>, sycl::event> compute_squared_l2_norms<F>( \
-        sycl::queue&,                                                       \
-        const ndview<F, 2>&,                                                \
-        const event_vector&,                                                \
-        const sycl::usm::alloc&);                                           \
-    template sycl::event scatter_2d<F>(sycl::queue & q,                     \
-                                       const ndview<F, 1>&,                 \
-                                       const ndview<F, 1>&,                 \
-                                       ndview<F, 2>&,                       \
-                                       const event_vector&);                \
-    template sycl::event compute_inner_product<F>(sycl::queue & q,          \
-                                                  const ndview<F, 2>&,      \
-                                                  const ndview<F, 2>&,      \
-                                                  ndview<F, 2>&,            \
-                                                  const event_vector&);     \
     template class distance<F, squared_l2_metric<F>>;
 
 INSTANTIATE(float);
